requirements.txt
camoufox
python-dotenv
requests
python-dateutil
tzdata
